test: Move recovery test to own test file
